New ID insertion
Last method of inserting an ID was using the same method as a normal insertion and letting auto_increment property (Database attribute) to take care of it. That made it so it had a greater number each time without minding deleted entries.
New method loops until the first free id is found.